+++ PgRoutiner/SettingsManagement/ParseSettings.cs
@@ -82,16 +82,31 @@ namespace PgRoutiner.SettingsManagement
 
                 if (!string.IsNullOrEmpty(Value.ConfigPath))
                 {
-                    Value.ConfigPath = Path.Join(Program.CurrentDir, Value.ConfigPath);
-                    if (File.Exists(settingsFile))
+                    configBuilder = new ConfigurationBuilder();
+                    pgroutinerFile = Path.Join(Value.ConfigPath, pgroutinerSettingsFile);
+                    if (File.Exists(pgroutinerFile))
                     {
-                        files.Add(" " + Path.GetFileName(Value.ConfigPath));
+                        files.Add(" " + Path.GetFileName(pgroutinerFile));
+                        configBuilder.AddJsonFile(pgroutinerFile, optional: true, reloadOnChange: false);
                     }
-                    configBuilder = new ConfigurationBuilder()
-                        .AddJsonFile(pgroutinerFile, optional: true, reloadOnChange: false)
-                        .AddJsonFile(settingsFile, optional: true, reloadOnChange: false)
-                        .AddJsonFile(devSettingsFile, optional: true, reloadOnChange: false)
-                        .AddJsonFile(Value.ConfigPath, optional: true, reloadOnChange: false);
+                    pgroutinerFile2 = Path.Join(Value.ConfigPath, "pgroutiner.json");
+                    if (File.Exists(pgroutinerFile2))
+                    {
+                        files.Add(" " + Path.GetFileName(pgroutinerFile2));
+                        configBuilder.AddJsonFile(pgroutinerFile2, optional: true, reloadOnChange: false);
+                    }
+                    settingsFile = Path.Join(Value.ConfigPath, "appsettings.json");
+                    if (File.Exists(pgroutinerFile2))
+                    {
+                        files.Add(" " + Path.GetFileName(settingsFile));
+                        configBuilder.AddJsonFile(devSettingsFile, optional: true, reloadOnChange: false);
+                    }
+                    devSettingsFile = Path.Join(Value.ConfigPath, "appsettings.Development.json");
+                    if (File.Exists(pgroutinerFile2))
+                    {
+                        files.Add(" " + Path.GetFileName(devSettingsFile));
+                        configBuilder.AddJsonFile(devSettingsFile, optional: true, reloadOnChange: false);
+                    }
                     config = configBuilder.Build();
                     config.GetSection("PgRoutiner").Bind(Value);
                     config.Bind(Value);